+++ src/main.c
@@ -14,7 +14,7 @@
 #include <sys/stat.h>
 #include <unistd.h>
 
-static constexpr char DIAMOND_VERSION[] = "0.3.0";
+static constexpr char DIAMOND_VERSION[] = "0.4.0";
 
 static char *read_file(const char *path) {
     FILE *file = fopen(path, "rb");